requirements.txt
python-telegram-bot[job-queue]
requests
GitPython
flask
